requirements.txt
uvicorn api:app --reloadfastapi
uvicorn
streamlit
opencv-python
faiss-cpu
numpy
torch
transformers
Pillow
